Revert "Preliminary support for regression models"
This reverts commit 9c4f1065f14f3be99bae64739ca2464ac893cbf1.

+++ selene_sdk/samplers/file_samplers/mat_file_sampler.py
@@ -225,5 +225,5 @@ class MatFileSampler(FileSampler):
         sequences_and_targets.append((seqs, tgts))
         targets_mat.append(tgts)
         # TODO: should not assume targets are always integers
-        targets_mat = np.vstack(targets_mat).astype(float)
+        targets_mat = np.vstack(targets_mat).astype(int)
         return sequences_and_targets, targets_mat
